Create FUNDING.yml
I love fishing and wreck hunting and love garmin products but really wanted this type of access for the files. I am not a professional coder and this has taken me a long time to figure out and parse the rsd files. Any tip you want to throw my way please do. Long term I hope to build installers for windows and Mac